create separate reducer ImportAccount
