+++ src/views/Wallet/views/Import/Container.js
@@ -2,7 +2,7 @@
 import * as React from 'react';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
-import * as AccountsAction from 'actions/AccountsActions';
+import * as ImportAccountActions from 'actions/ImportAccountActions';
 
 import type { MapStateToProps, MapDispatchToProps } from 'react-redux';
 import type { TrezorDevice, Config, State, Dispatch } from 'flowtype';
@@ -10,12 +10,13 @@ import ImportView from './index';
 
 export type StateProps = {
     device: ?TrezorDevice,
+    importAccount: $ElementType<State, 'importAccount'>,
     config: Config,
     children?: React.Node,
 };
 
 type DispatchProps = {
-    importAddress: typeof AccountsAction.importAddress,
+    importAddress: typeof ImportAccountActions.importAddress,
 };
 
 type OwnProps = {};
@@ -27,12 +28,13 @@ const mapStateToProps: MapStateToProps<State, OwnProps, StateProps> = (
 ): StateProps => ({
     config: state.localStorage.config,
     device: state.wallet.selectedDevice,
+    importAccount: state.importAccount,
 });
 
 const mapDispatchToProps: MapDispatchToProps<Dispatch, OwnProps, DispatchProps> = (
     dispatch: Dispatch
 ): DispatchProps => ({
-    importAddress: bindActionCreators(AccountsAction.importAddress, dispatch),
+    importAddress: bindActionCreators(ImportAccountActions.importAddress, dispatch),
 });
 
 export default connect(
